--- seqs.sql
@@ -1,20 +1,0 @@
-CREATE SEQUENCE IF NOT EXISTS championship_seq
-    START with 1
-    INCREMENT BY 1
-    NO MINVALUE
-    NO MAXVALUE
-    CACHE 1;
-
-CREATE SEQUENCE IF NOT EXISTS teams_seq
-    START with 1
-    INCREMENT BY 1
-    NO MINVALUE
-    NO MAXVALUE
-    CACHE 1;
-
-CREATE SEQUENCE IF NOT EXISTS players_seq
-    START with 1
-    INCREMENT BY 1
-    NO MINVALUE
-    NO MAXVALUE
-    CACHE 1;
