--- src/main/java/br/com/nayaranunes/terradonunca/domain/Champions.java
@@ -1,11 +1,0 @@
-package br.com.nayaranunes.terradonunca.domain;
-
-import java.util.List;
-
-public class Champions {
-    private List<Integer> champions;
-
-    public List<Integer> getChampions() {
-        return champions;
-    }
-}
