--- src/main/java/br/com/nayaranunes/terradonunca/exception/ApiRequestException.java
@@ -1,13 +1,0 @@
-package br.com.nayaranunes.terradonunca.exception;
-
-public class ApiRequestException extends RuntimeException {
-    //costumer exception
-
-    public ApiRequestException(String message) {
-        super(message);
-    }
-
-    public ApiRequestException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
