--- src/test/java/br/com/nayaranunes/terradonunca/TerradonuncaApplicationTests.java
@@ -1,21 +1,0 @@
-package br.com.nayaranunes.terradonunca;
-
-import org.junit.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-public class TerradonuncaApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-	/*
-		A test has three steps:
-		1. scenario
-		2. action
-		3. verification
-	*/
-
-}
-
