+++ tdn-common/src/main/java/br/com/nayaranunes/terradonunca/ApiException.java
@@ -1,0 +1,30 @@
+package br.com.nayaranunes.terradonunca;
+
+import org.springframework.http.HttpStatus;
+
+import java.time.ZonedDateTime;
+
+public class ApiException {
+    private final String message;
+    //private final Throwable throwable;
+    private final HttpStatus httpStatus;
+    private final ZonedDateTime timestamp;
+
+    public ApiException(String message, HttpStatus httpStatus, ZonedDateTime timestamp) {
+        this.message = message;
+        this.httpStatus = httpStatus;
+        this.timestamp = timestamp;
+    }
+
+    public String getMessage() {
+        return message;
+    }
+
+    public HttpStatus getHttpStatus() {
+        return httpStatus;
+    }
+
+    public ZonedDateTime getTimestamp() {
+        return timestamp;
+    }
+}
