+++ tdn-common/src/main/java/br/com/nayaranunes/terradonunca/ApiRequestException.java
@@ -1,0 +1,13 @@
+package br.com.nayaranunes.terradonunca;
+
+public class ApiRequestException extends RuntimeException {
+    //costumer exception
+
+    public ApiRequestException(String message) {
+        super(message);
+    }
+
+    public ApiRequestException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
