+++ tdn-data-acess/src/main/java/br/com/nayaranunes/terradonunca/entitys/Champions.java
@@ -1,0 +1,11 @@
+package br.com.nayaranunes.terradonunca.entitys;
+
+import java.util.List;
+
+public class Champions {
+    private List<Integer> champions;
+
+    public List<Integer> getChampions() {
+        return champions;
+    }
+}
